requirements.txt
flask
flask_sqlalchemy
python-barcode
Pillow
python-dateutil